Updated Entity Definitions to Match the Style Guide
The Style Guide indicated that embedded/reference Entities should be named with a prefix of "r" followed by the Entity name.  I updated all of the Consumer Credit Entities to follow that naming convention as well as the various black box Entities to utilize "Person rPerson" instead of "Person Subject".