Bankruptcy Shell Changes and First Degree Bankruptcy Count
Now that 0.8.2 has released it fixes the previously seen issue with
First Degree Relatives and Associates Bankruptcy Count causing the
compiler to lock up and spin.  I am uncommenting that attribute - note
that because of this the latest 0.8.2 KEL Gold release is required.

I am also updating the Consumer SHELL to output Bankruptcy Disposition
instead of Bankruptcy Status, as Disposition contains the values that
are being output in the RAMPS Consumer Shell Service.