Initial CLUE Black Box Development - Base Files
Added ECL code to clean and normalize the CLUE Auto and CLUE Property base files.

As part of this process I am currently appending a unique record identifier, this will eventually be burned into the base files.  But everything else will continue to be done during the KEL ingest process.