CLUE Property ENTITY Development
The CLUE Property Black Box ENTITY development is complete and tested. A total of 4 ENTITY's were created as part of this effort:
CLUEProperty, CLUEPropertySearchInformation, CLUEPropertyRiskClaims, CLUEPropertyInsuredClaims

In total all 4 entities contain roughly 280 fields that are all populated in the base data.  Note that any fields not populated (confirmed with SALT profiles) were excluded from these entity definitions to keep them as small and manageable as possible.