Quick LexID Filter
Added the ability to filter all of the LexID based files by modifying a single location to include a set of LexIDs that we want to include in our KEL Base.  This filters all of the base files to only keep around those LexID's resulting in much faster Persist file generation during KEL code development and initial Unit Testing.  Once initial development and testing is complete, this filtering dataset can be set to a blank dataset to run the entire file.

The LexID filtering set can be found in the new attribute: _Control/LexIDFilterSet.ecl