DAYS/MONTHS/YEARSBETWEEN Absolute Value Fixes
It was recently discovered that when using DAYSBETWEEN/MONTHSBETWEEN/YEARSBETWEEN you could get negative values if the larger date ended up on the left-hand side of the function call.  This presents problems when performing DAYSBETWEEN(date1, date2) <= 30 checks to only count records seen in the last 30 days as you could have a record that returns -45 and is suddenly counted, even though the two dates are really 45 days apart.

Also flipping to using CURRENTDATE() instead of TODAY(), as CURRENTDATE() is ERA/ASOF compliant, while TODAY() always returns the date the code was executed.