NCF v2.0 Credit Report - Initial Code
Have gone through and cleaned up the NCF v2.0 Credit Report ENTITY model to better match our coding standards and best practices.  This is the initial version of that code, and could potentially still go through several revisions.  Committing to the main project to begin work on the requested CONSTANTS and Global Attributes.