Product Results ENTITY Cleanup
Renamed CLUEAuto to ClaimsAuto and included SHELL's for CLUEAuto and Claims Discovery Auto.
Renamed CLUEProperty to ClaimsProperty and included SHELL's for CLUEProperty and Claims Discovery Property.
Renamed CarrierAuto to PolicyAuto and included SHELL's for Carrier Discovery Auto and Current Carrier Auto.
Renamed CarrierProperty to PolicyProperty and included a SHELL for Current Carrier Property.

Updated the ECL code for Claims Discovery Auto and Claims Discovery Policy so that the updated CLUE Payments structure was compatible with the Claims Discovery Cleaned ECL code.