Added the Email Entity to Graph.kel
I added the Email Entity to Graph.kel and added the requisite import ecl files in Email_Data. I also added the PersonEmail association as well as some basic attributes as described  in the ecl code for bocashell. There are also three simple queries to test each, the new entity, the association, and then the new attributes to person.