KEL Code Restructure
KnowledgeUniverse will be brought into the various repositories (Vault, PublicRecords, Alpha) as a Git Sub-Module folder.

To support this, we need to update the Attributes.kel and Graph.kel to have the proper IMPORT syntax.

As part of this also fixing errors with DriversLicense ENTITY and Drivers ASSOCIATIONs (It would appear we had old placeholder code that wasn't removed).  For this ticket that involves updating the ENTITYs and ASSOCIATIONs in Entities.kel.