KS-1112 MultiValued Properties - FCRA Name Collapse
Per QA of KS-316 we are collapsing the FCRATrafficFlag into the TrafficFlag field, and the FCRAConvictionFlag into the ConvictionFlag field.

Per QA of KS-1112 it was discovered that several properties contain multiple values, and it was randomly selecting a single value.  Converting those properties to multi-valued properties:
CriminalOffense:
CourtOffenseLevel
OffenderLevel

CriminalOffender:
NumberOfOffenseCounts