Initial Version - GitLab CI YML for Generating a New Tag Daily
Initial YML file for integrating the Vault repository with the Innovation Cloud Shared runners.  This initial file simply generates new Git tags every night.

Everything in the before_script section sets up Git in the Docker Container running on the Innovation Cloud (As that is currently missing from the Docker image as part of the Innovation Cloud shared runner).

Inside of the script I am performing a shallow clone of the repository (Only grabbing the most recent commit from the master branch), adding origin with the proper Username/Password (Stored within the Project as a protected keypair), and then pushing the new tag back up to the project.

Only specifies this particular YML file will only execute within the scheduled pipeline (Set for 11PM daily)

Tags specifies information about which shared runner(s) are allowed to execute this job, those 4 tags are specified for the Innovation Cloud Shared runners A and B.