KS-2861 Update SeleTradelines
Per KS-2861 we are removing the Business ENTITY and instead directing everything to the new BusinessSele ENTITY. The UID components for both entities was identical, thus is was mostly just renamed Business references to BusinessSele.

As part of this change, the Analytic Review of the BusinessSele ENTITY requested a more standardized naming convention to help distinguish what level of a Business the Association happens at, thus BusinessTradeline was renamed to SeleTradeline (for example).

Lastly, I have made some performance enhancements to the MAC_Determine_Multi_Value script, this should help avoid memory limit exceeded errors, and if necessary allows for ignoring entire sub-models.