Inquiry UID Fix - Uniqueness
We have discovered that TransactionID alone is not enough to uniquely identify and Inquiry record. Our Batch systems appear to run an entire customer file with the same TransactionID, and then upload that information to our Inquiry database. They do however increment a SequenceNumber field when processing that customer file.

Because of this we are adapting the Inquiry ENTITY to have a UID of TransactionID + SequenceNumber. This drops the highest UID field count from > 2,000,000 records to 20 records. Those records appear to be all the same transaction, just with different LexID's appended.