Index Building Temporary QUERY
I have added a new QUERY which is intended to contain the Index filters that are temporarily needed to perform some sort of testing. If you add a new filter here, it will be picked up by the Index Automation process and automatically build your new lookup table. When you no longer need this index, delete it from the Temp QUERY and push to master - it will no longer be built by the Index Automation.