KS-3235 Remove SelePerson Update Business Attributes
Per KS-3235 we are removing the SelePerson ASSOCIATION, and updating the logic for a handful of Business Attributes to use Header data only. This merge request removes the ASSOCIATION definition, and updates the Attribute logic.

Per KS-4769 we are also enhancing the QUERY code to ensure more association indexes we need are built. This merge request includes the updated QUERY code for those additional indexes.