KS-4178: LienJudgment ENTITY Review Updates
Per Analytic Review we are renaming the EvictionFlag field, and removing references to Release Date based on feedback from Legal.  Technically the eviction flag doesn't indicate if the person was actually evicted, just that the Landlord has requested/started the process of eviction. To be favorable to consumers, we are renaming to avoid that misconception.  Additionally Release Date is very, very poorly updated by our data vendors (It actually requires that we physically send someone to pull the court documents, and then manually update the data), and as such Legal has requested we stop using it completely.  This will mean that DateLastSeen is no longer populated for the LienJudgment ENTITY.