Updated the ClaimsCompassECrash_KEL Attributes to Avoid ALL JOIN
Updated the ClaimsCompassEcrash_KEL/Attributes.kel to avoid an ALL JOIN that was being generated on the Vehicle, Accident, and VehicleAccident Indexes.  The primary issue was around the topnhelper Attributes - the KEL Compiler wasn't properly identifying an optimization that needed to happen. A JIRA ticket has been opened to resolve this in the future.