Vehicle Master File Enhancements - Combine QUERY, Include Attribute Stubs
Updated the QUERY code to merge the two QUERY's into one per request from Data Engineering. Also stubbed in the list of Attributes that are expected to be returned from each QUERY.

Fixed the VehicleGroupID having a space in the key value.

Added a spot for ValidVIN to be computed as part of the FileCleaning process - will need to replace with the correct cleaning function later on.