CurrentCarrier_KEL - Add InceptionDateBatch
Added InceptionDateBatch to the Policy Entity to help hold the new Batch data which isn't a raw date but instead a bucketed date string value such as < 6 Months or > 24 Months.

Created a new CrossProductReport ENTITY which links together Reports with the same TransactionID but differing Products (Such as Batch Risk Address Search and Subject Search).